backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
requests>=2.31,<3.0
